// plants-finder/src/app/interfaces/plant.ts
export interface IPlant {
  latin_name: string;
  common_name: string;
  family_name: string;
  plant_type: string;
  bloom_time: string;
  size_at_maturity: string;
  climate_appropriate_plants: string;
  suitable_site_conditions: string;
  water_needs: string;
  stormwater_benefit: string;
  stormwater_int: string;
  appropriate_location: string;
  suggested_green_connection_routes: string;
  street_tree_list: string;
  additional_characteristices_notes: string;
  photocredit01: string;
  photocredit02: string;
  photocredit03: string;
  photocredit04: string;
  sidewalk_landscaping_plants: string;
  sidewalk_landscaping_plants_int: number;
}